- moved getURL methods into S3Service - relocate link creation to where other links are created
